refactor: extract MapUpdater to shared MapHelperComponents
- Move MapUpdater from SiteDetailView to MapHelperComponents.tsx
- Update SiteDetailView to import shared MapUpdater component
- Update test mocks to include MapUpdater
- Reduces code duplication (12 lines of duplicated code eliminated)
- Improves maintainability and consistency

All 194 tests passing.
